+++ Libplanet/Action/IAccountStateView.cs
@@ -8,11 +8,15 @@ namespace Libplanet.Action
     public interface IAccountStateView
     {
         /// <summary>
-        /// Gets an account state of the given <see cref="Address"/>.
+        /// Gets an account state of the given <paramref name="address"/>.
+        /// <para>If the given <paramref name="address"/> has never been set
+        /// its account status, returns <c>null</c> instead of throwing
+        /// any exception.</para>
         /// </summary>
         /// <param name="address">An address of the account to read
         /// its state.</param>
-        /// <returns>The account state.</returns>
+        /// <returns>The account state if exists.  Otherwise <c>null</c>.
+        /// </returns>
         object GetAccountState(Address address);
     }
 }