+++ CHANGES.md
@@ -12,6 +12,9 @@ To be released.
     methods were also renamed too.
  -  Added `IAccountStateView` interface, and made `BlockChain<T>` implements it.
     The interface purposes to provide a read-only view to account states.
+ -  Added `IAccountStateDelta` interface.
+    The interface purposes to provide a read-write view to account states
+    with maintaining `UpdatedAddresses` (so-called "dirty").
  -  The type of `Peer.Urls` property was changed from `Uri` to `IPEndPoint`.
  -  Since we decided to depend on TURN ([RFC 5766]) and STUN ([RFC 5389]) to
     work around NAT so that `Peer`'s endpoints don't have to be multiple,